+++ tests/testthat/test-assertions.R
@@ -105,3 +105,83 @@ test_that("assert_PKNCAdata", {
     regexp = "Must be a PKNCAdata object"
   )
 })
+
+test_that("element_find", {
+  values5<-c(TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, TRUE)
+  values10<-c(TRUE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, TRUE)
+  expect_equal(
+    element_find(values5),
+    element_find(values10)
+  )
+  
+  expect_equal(
+    element_find(values5),
+    "Elements 1, 2, 3, 4, 5"
+  )
+  
+  expect_equal(
+    element_find(values10),
+    "Elements 1, 2, 3, 4, 5"
+  )
+})
+
+test_that("assert_unit_col", {
+  d <- data.frame(A = 1, B = "C")
+  expect_equal(
+    assert_unit_col(unit = "B", data = d),
+    structure("B", unit_type = "column")
+  )
+  expect_error(
+    assert_unit_col(unit = 1:2),
+    regexp = "`unit` must be a single value"
+  )
+  expect_error(
+    assert_unit_col(unit = 1),
+    regexp = "`unit` must be a character string"
+  )
+  expect_error(
+    assert_unit_col(unit = "D", data = "A"),
+    regexp = "`data` must be a data.frame"
+  )
+  expect_error(
+    assert_unit_col(unit = "D", data = d),
+    regexp = "`unit` (D) must be a column name in the data",
+    fixed = TRUE
+  )
+  expect_error(
+    assert_unit_col(unit = "A", data = d),
+    regexp = "`unit` (A) must contain character data",
+    fixed = TRUE
+  )
+})
+
+test_that("assert_unit_value", {
+  expect_equal(assert_unit_value("A"), structure("A", unit_type = "value"))
+  expect_null(assert_unit_value(NULL))
+  expect_error(
+    assert_unit_value(c("A", "B")),
+    regexp = "`unit` must be a single value"
+  )
+  expect_error(
+    assert_unit_value(1),
+    regexp = "`unit` must be a character string"
+  )
+})
+
+test_that("assert_unit", {
+  d <- data.frame(A = 1, B = "C")
+
+  expect_null(assert_unit(NULL, data = d))
+  expect_equal(
+    assert_unit(unit = "B", data = d),
+    structure("B", unit_type = "column")
+  )
+  expect_equal(
+    assert_unit(unit = "C", data = d),
+    structure("C", unit_type = "value")
+  )
+  expect_error(
+    assert_unit(unit = 1, data = d),
+    regexp = "`unit` must be a character string"
+  )
+})